Rename: 업로드 페이지 컴포넌트 모두 post에서 upload로 이름변경
